% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download LawLibrary.AI data}
\usage{
download_data(database_id, dataset_id, api_key, parameters = NULL)
}
\arguments{
\item{database_id}{A string. The ID of a database. Run \code{get_databases()}
to get a list of valid values.}

\item{dataset_id}{A string. The ID of a dataset in the specified database.
Run \code{get_databases()} to get a list of valid values.}

\item{api_key}{A string. Your LawLibrary.AI API key.}

\item{parameters}{A named list. The default is \code{NULL}. Each element in
the list specifies a parameter to apply to the dataset. Only some variables
in each dataset can be used to filter. See the documentation for the API
for more information. If you want to filter by a variable that isn't
supported by the API, use \code{dplyr::filter()} after downloading the
data. Each element should be the name of a variable in the specified
dataset and the corresponding value should be a value or vector of values
that the variable can take. The response will only include observations
where the variable equals one of the provided values. If you specify
multiple parameters, the results will only include observations that match
all of the parameters. For some variables, you can add \code{_min} or
\code{_max} to the end of the variable name to specify a minimum or maximum
value. The API will ignore invalid parameters.}
}
\value{
This function returns a tibble containing the requested data.
}
\description{
This function allows you to download data from LawLibrary.AI
  databases via the LawLibrary.AI API. You need to choose a database and a
  dataset in that database to download. You can specify filters to select
  specific observations.
}
\details{
The LawLibrary.AI API has a rate limit. This function automatically
  manages compliance with the rate limit by downloading data from the API in
  batches. It downloads 10,000 observations every 5 seconds. It prints a
  message to the console that indicates how many observations you have
  requested and approximately how long it will take to download the data. It
  also prints a message after every batch that indicates the current
  progress. After your download is complete, it will print the suggested
  citation for the data.
}
\examples{
\dontrun{
data <- download_data(
  database_id = "state_aid",
  dataset_id = "cases",
  parameters = list(
    min_date = "2010-01-01",
    max_date = "2019-12-31",
    member_state = c("France", "Germany")
  ),
  api_key = "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
)}

}
