% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_citation.R
\name{get_citation}
\alias{get_citation}
\title{Print the citation for LawLibrary.AI database}
\usage{
get_citation(database_id, api_key)
}
\arguments{
\item{database_id}{A string. The ID for the database.}

\item{api_key}{A string. Your LawLibrary.AI API key.}
}
\description{
This function generates the recommended citation for a LawLibrary.AI
database. You need to provide the ID for the database. You can run
\code{get_databases()} to get a list of valid values.
}
\examples{
\dontrun{
print_citation(
  database_id = "state_aid",
  api_key = "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
)}

}
