% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_databases.R
\name{get_databases}
\alias{get_databases}
\title{Get a summary of all LawLibrary.AI databases}
\usage{
get_databases(api_key)
}
\arguments{
\item{api_key}{A string. Your LawLibrary.AI API key.}
}
\description{
This function returns a table that summarizes all LawLibrary.AI databases.
The table indicates all of the databases that are currently available and all
of the datasets that are currently available in each database. The table
includes an ID for each database and dataset, which you'll need to provide to
\code{download_data()} in order to download the data.
}
\examples{
\dontrun{
get_databases(
  api_key = "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
)}

}
